pyproject.toml
[project]
name = "predictivebio"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["core*", "agents*", "tools*", "services*", "db*"]
